requirements.txt
fastapi
uvicorn
fastlab
pytest
requests
googleapiclient
jinja2
numpy
Pillow
python-multipart
matplotlib